fix(grpc_client): fix bug where it would have this delayed crashing
delayed crashing came from the fact that main loop was not connected to
the market. since threads do not depend on each other, the main loop
thread still ran even though a market server was not valid. since market
server is not connected and that there is nothing that tells the main
loop thread that there is no command executor nor market server
connected, the main loop thread still ran and only crashed when it would
deal with input. this commit uses a condvar to notify the main loop to
crash.